UI: Rename sign[in/up] into registration/login
