+++ monkey/monkey_island/cc/next_ui/src/app/(auth)/login/useRedirectToRegistration.tsx
@@ -2,9 +2,8 @@ import { useRouter } from 'next/navigation';
 import { useRegistrationStatusQuery } from '@/redux/features/api/authentication/authenticationEndpoints';
 import { useEffect } from 'react';
 import { PATHS } from '@/constants/paths.constants';
-import { tokenIsStored } from '@/lib/authenticationToken';
 
-const useRedirectToLogin = () => {
+const useRedirectToRegistration = () => {
     const router = useRouter();
     const { data: registrationStatus, isLoading: isRegistrationStatusLoading } =
         useRegistrationStatusQuery();
@@ -12,12 +11,11 @@ const useRedirectToLogin = () => {
     useEffect(() => {
         if (
             !isRegistrationStatusLoading &&
-            !registrationStatus?.registrationNeeded &&
-            !tokenIsStored()
+            registrationStatus?.registrationNeeded
         ) {
-            router.push(PATHS.SIGN_IN);
+            router.push(PATHS.REGISTRATION);
         }
     }, [isRegistrationStatusLoading, registrationStatus, router]);
 };
 
-export default useRedirectToLogin;
+export default useRedirectToRegistration;
